setup add/edit screen

+++ app/src/main/java/com/fyp/inotes/feature_note/domain/model/Note.kt
@@ -13,7 +13,7 @@ data class Note(
     @PrimaryKey val id: Int? = null,
     val title: String,
     val content: String,
-    val color: Long,
+    val color: Int,
     val timeStamp: Long
 ) {
     companion object {
